Split up refactored dispatcher
