+++ lib/actions.js
@@ -1,0 +1,3 @@
+module.exports = {
+  comment: require('./actions/comment')
+};
